+++ packages/react/src/components/DropBox/DropBox.styles.js
@@ -1,6 +1,6 @@
 import { css } from '@emotion/react';
 
-export const dropBoxStyles = {
+const dropBoxStyles = {
    dropBoxCss: css`
       z-index: 300;
       opacity: 25%;
@@ -13,14 +13,16 @@ export const dropBoxStyles = {
       left: 0;
       width: 100%;
       height: 100%;
-      pointer-events: none; /* to allow interactions with the components underneath */
+      pointer-events: none; 
 `,
 
    borderCss: css`
       position: relative;
       top: 0;
       left: 0;
-      z-index: 9999; /* Choose a high value for z-index */
+      z-index: 9999; 
       border: 4px dashed #007fff !important;
   `,
 };
+
+export default dropBoxStyles;
