+++ packages/react/src/components/DropBox/index.js
@@ -1,0 +1,2 @@
+export { default as DropBoxOverlay } from './DropBoxOverlay';
+export { default as dropBoxStyles } from './DropBox.styles';
